+++ src/com/jmex/terrain/util/FaultFractalHeightMap.java
@@ -153,7 +153,7 @@ public class FaultFractalHeightMap extends AbstractHeightMap {
 		}
 
 		//allocate new arrays
-		heightData = new int[size*size];
+		heightData = new float[size*size];
 		float[][] tempBuffer = new float[size][size];
 
 		//generate faults for the number of iterations given.
@@ -198,7 +198,7 @@ public class FaultFractalHeightMap extends AbstractHeightMap {
 		//transfer the new terrain into the height map.
 		for (int i = 0; i < size; i++) {
 			for (int j = 0; j < size; j++) {
-				setHeightAtPoint((int) tempBuffer[i][j], j, i);
+				setHeightAtPoint((float) tempBuffer[i][j], j, i);
 			}
 		}
